+++ CHANGES.md
@@ -13,6 +13,7 @@
   - tasks are configured before project evaluation, `project.afterEvaluate` is not require anymore
 - fix incremental compiling after instrumenting code (fixes [#116](../../issues/116))
 - added `intellij.ideaDependencyCachePath` option (fixes [#127](../../issues/127))
+- `project()` reference can be used as a plugin-dependency (fixes [#17](../../issues/17))
 
 ## 0.1